requirements.txt
django
django-jalali-date
pillow
